backup_comprehensive_gap_analysis_20250711_002712/backup_phase3_comprehensive_20250711_000719/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
flask-socketio>=5.3.6
eventlet>=0.33.3
flask>=3.0.0
flask-cors>=4.0.0
bcrypt>=4.0.0
redis>=5.0.0
sqlalchemy>=2.0.23
alembic>=1.13.0
scikit-learn>=1.3.0
openai>=1.3.0
websockets>=12.0
speechrecognition>=3.10.0
anthropic>=0.34.0
structlog>=24.1.0
prometheus-client>=0.20.0
pydub>=0.25.1
textblob>=0.19.0
torch>=2.0.0
# Phase 2.1: Advanced Redis & Caching
hiredis>=3.0.0
# Phase 2.1: Performance Monitoring
psutil>=6.0.0
# Phase 2.1: Data Governance
pydantic-settings>=2.0.0
# Phase 2.1: Security Hardening
argon2-cffi>=23.1.0
# Phase 2.2: CDN & Analytics
cloudflare>=2.19.0
mixpanel>=4.10.0
# Phase 2.3: AI/ML Enhancements
transformers>=4.36.0
sentence-transformers>=2.2.2
langchain>=0.1.0
google-generativeai>=0.3.0
networkx>=3.0
# Phase 2.4: Push Notifications
firebase-admin>=6.4.0
